Adicionando Pessoa e alterando Passageiro e Motorista
