+++ SistemaTransporte/src/modelo/Motorista.java
@@ -8,30 +8,19 @@ package modelo;
  *
  * @author kauan
  */
-public class Motorista {
+public class Motorista extends Pessoa{
  
-    private String nome;
     private String cnh; 
-    private int idade;
     private int experiencia; 
     private String empresa;  
     private String linhaOnibus;
 
     public Motorista(String nome, String cnh, int idade, int experiencia, String empresa, String linhaOnibus) {
-        this.nome = nome;
+        super(nome, idade);
         this.cnh = cnh;
-        this.idade = idade;
         this.experiencia = experiencia;
         this.empresa = empresa;
         this.linhaOnibus = linhaOnibus;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
     }
 
     public String getCnh() {
@@ -40,14 +29,6 @@ public class Motorista {
 
     public void setCnh(String cnh) {
         this.cnh = cnh;
-    }
-
-    public int getIdade() {
-        return idade;
-    }
-
-    public void setIdade(int idade) {
-        this.idade = idade;
     }
 
     public int getExperiencia() {
@@ -74,11 +55,10 @@ public class Motorista {
         this.linhaOnibus = linhaOnibus;
     }
 
-    
+    @Override
     public void exibirInfo() {
-        System.out.println("Nome do Motorista: " + nome);
+        super.exibirInfo();
         System.out.println("CNH: " + cnh);
-        System.out.println("Idade: " + idade + " anos");
         System.out.println("Experiência de: " + experiencia + " anos dirigindo");
         System.out.println("Empresa: " + empresa);
         System.out.println("Faz a linha: " + linhaOnibus);
