+++ SistemaTransporte/src/modelo/Passageiro.java
@@ -1,37 +1,18 @@
 package modelo;
 
-public class Passageiro implements Cloneable{
+public class Passageiro extends Pessoa implements Cloneable{
     //atributos
-    private String nome;
-    private int idade;
     private int telefone;
     private String cpf;
 
     //método construtor
     public Passageiro(String nome, int idade, int telefone, String cpf) {
-        this.nome = nome;
-        this.idade = idade;
+        super(nome, idade);
         this.telefone = telefone;
         this.cpf = cpf;
     }
     
-    //métodos getters e setters dos atributos nome, idade, telefone e cpf do passageiro
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public int getIdade() {
-        return idade;
-    }
-
-    public void setIdade(int idade) {
-        this.idade = idade;
-    }
-    
+    //métodos getters e setters
     public void setTelefone(int telefone){
         this.telefone = telefone;
     }
@@ -48,7 +29,14 @@ public class Passageiro implements Cloneable{
         this.cpf = cpf;
     }
     
-     @Override
+    @Override
+    public void exibirInfo() {
+        super.exibirInfo();
+        System.out.println("Telefone: " + telefone);
+        System.out.println("CPF: " + cpf);
+    }
+    
+    @Override
     public Passageiro clone(){
         Passageiro passageiro;
         try{
