+++ UMLs/Passageiro.drawio
@@ -1,17 +1,17 @@
-<mxfile host="app.diagrams.net" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36" version="24.7.12">
+<mxfile host="app.diagrams.net" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36" version="24.8.3">
   <diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">
-    <mxGraphModel dx="989" dy="538" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="827" pageHeight="1169" math="0" shadow="0">
+    <mxGraphModel dx="1434" dy="780" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="827" pageHeight="1169" math="0" shadow="0">
       <root>
         <mxCell id="WIyWlLk6GJQsqaUBKTNV-0" />
         <mxCell id="WIyWlLk6GJQsqaUBKTNV-1" parent="WIyWlLk6GJQsqaUBKTNV-0" />
         <mxCell id="UeuIvZKMgW835NmC7W0M-0" value="Passageiro" style="swimlane;fontStyle=1;align=center;verticalAlign=top;childLayout=stackLayout;horizontal=1;startSize=26;horizontalStack=0;resizeParent=1;resizeParentMax=0;resizeLast=0;collapsible=1;marginBottom=0;whiteSpace=wrap;html=1;fillColor=#fff2cc;strokeColor=#d6b656;" parent="WIyWlLk6GJQsqaUBKTNV-1" vertex="1">
-          <mxGeometry x="289" y="190" width="250" height="124" as="geometry" />
+          <mxGeometry x="289" y="190" width="250" height="110" as="geometry" />
         </mxCell>
-        <mxCell id="UeuIvZKMgW835NmC7W0M-1" value="- nome: String;&lt;div&gt;- idade: int;&lt;/div&gt;&lt;div&gt;- telefone: int;&lt;br&gt;&lt;div&gt;- cpf: String;&lt;/div&gt;&lt;/div&gt;" style="text;strokeColor=#d6b656;fillColor=#fff2cc;align=left;verticalAlign=top;spacingLeft=4;spacingRight=4;overflow=hidden;rotatable=0;points=[[0,0.5],[1,0.5]];portConstraint=eastwest;whiteSpace=wrap;html=1;" parent="UeuIvZKMgW835NmC7W0M-0" vertex="1">
-          <mxGeometry y="26" width="250" height="74" as="geometry" />
+        <mxCell id="UeuIvZKMgW835NmC7W0M-1" value="&lt;span style=&quot;background-color: initial;&quot;&gt;- telefone: int;&lt;/span&gt;&lt;div&gt;&lt;div&gt;- cpf: String;&lt;/div&gt;&lt;/div&gt;" style="text;strokeColor=#d6b656;fillColor=#fff2cc;align=left;verticalAlign=top;spacingLeft=4;spacingRight=4;overflow=hidden;rotatable=0;points=[[0,0.5],[1,0.5]];portConstraint=eastwest;whiteSpace=wrap;html=1;" parent="UeuIvZKMgW835NmC7W0M-0" vertex="1">
+          <mxGeometry y="26" width="250" height="44" as="geometry" />
         </mxCell>
-        <mxCell id="UeuIvZKMgW835NmC7W0M-3" value="+ clone(): Passageiro;" style="text;strokeColor=#d6b656;fillColor=#fff2cc;align=left;verticalAlign=top;spacingLeft=4;spacingRight=4;overflow=hidden;rotatable=0;points=[[0,0.5],[1,0.5]];portConstraint=eastwest;whiteSpace=wrap;html=1;" parent="UeuIvZKMgW835NmC7W0M-0" vertex="1">
-          <mxGeometry y="100" width="250" height="24" as="geometry" />
+        <mxCell id="UeuIvZKMgW835NmC7W0M-3" value="+ clone(): Passageiro;&lt;div&gt;+ exibirInfo(): void;&lt;/div&gt;" style="text;strokeColor=#d6b656;fillColor=#fff2cc;align=left;verticalAlign=top;spacingLeft=4;spacingRight=4;overflow=hidden;rotatable=0;points=[[0,0.5],[1,0.5]];portConstraint=eastwest;whiteSpace=wrap;html=1;" parent="UeuIvZKMgW835NmC7W0M-0" vertex="1">
+          <mxGeometry y="70" width="250" height="40" as="geometry" />
         </mxCell>
       </root>
     </mxGraphModel>
